Move whenTrue into ui-logic
