+++ packages/sanes-chrome-extension/src/utils/test/navigation.ts
@@ -1,5 +1,6 @@
 import TestUtils from "react-dom/test-utils";
 import { Store } from "redux";
+import { whenTrue } from "ui-logic";
 
 import { GetPersonaResponse } from "../../extension/background/model/backgroundscript";
 import { Request } from "../../extension/background/model/signingServer/requestQueueManager";
@@ -17,33 +18,6 @@ import {
 import { REQUEST_FIELD } from "../../routes/requests/components/RequestList";
 import { history } from "../../store/reducers";
 import { createDom } from "../../utils/test/dom";
-
-const retryInterval = 400;
-
-/**
- * Calls callback until it returns true or timeout is reachend.
- *
- * @param callback a callback function that is called in a loop. It should be cheap to execute.
- * @param timeout timeout in milliseconds. Defaults to 10 seconds
- * @returns a promise that resolves when the callback returned true and is rejected when timeout is reached.
- */
-export function whenTrue(callback: () => boolean | Promise<boolean>, timeout: number = 10000): Promise<void> {
-  return new Promise((resolve, reject): void => {
-    const startTime = Date.now();
-    const interval = setInterval(async () => {
-      const runtime = Date.now() - startTime;
-      if (runtime >= timeout) {
-        clearInterval(interval);
-        reject(`Timeout reached after ${runtime} ms`);
-      } else {
-        if (await callback()) {
-          clearInterval(interval);
-          resolve();
-        }
-      }
-    }, retryInterval);
-  });
-}
 
 export const whenOnNavigatedToRoute = (refreshStore: Store, desiredRoute: string): Promise<void> =>
   whenTrue(() => refreshStore.getState().router.location.pathname === desiredRoute);
