+++ packages/ui-logic/types/index.d.ts
@@ -1,2 +1,3 @@
 export { singleton } from "./singleton";
 export { ellipsify } from "./strings";
+export { whenTrue } from "./testing";
